rename to literature_values
