+++ flowsa/data_source_scripts/EIA_CBECS_Land.py
@@ -13,7 +13,7 @@ import numpy as np
 from flowsa.common import US_FIPS, get_region_and_division_codes, WITHDRAWN_KEYWORD,\
     clean_str_and_capitalize, fba_mapped_default_grouping_fields, vLogDetailed
 from flowsa.flowbyfunctions import assign_fips_location_system, aggregator
-from flowsa.values_from_literature import \
+from flowsa.literature_values import \
     get_commercial_and_manufacturing_floorspace_to_land_area_ratio
 from flowsa.validation import calculate_flowamount_diff_between_dfs
